python/requirements.txt
kokoro-tts
soundfile
torch
Flask
